Split into per-component files & user weight logic
Next up: Switch to Typescript because type sigs would be helpful.
Once Typescript is satisfied, add tests for user weight logic.
